# specs/005-devexpress-reports/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: DevExpress Reports API
  description: |
    API contracts for the DevExpress Reports integration in Herbal Medicine ERP.

    This specification covers two API surfaces:
    1. **Next.js API** - Report management, permissions, and data endpoints
    2. **ASP.NET Core API** - DevExpress reporting engine endpoints

    Authentication uses JWT tokens passed in the Authorization header.
  version: 1.0.0
  contact:
    name: Herbal Medicine ERP
servers:
  - url: http://localhost:3000/api
    description: Next.js API (Development)
  - url: http://localhost:5000
    description: ASP.NET Core Reporting Backend (Development)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from ERP authentication
  schemas:
...
    ReportCategory:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
        name:
          type: string
          maxLength: 100
          description: Category display name
        description:
          type: string
          maxLength: 500
          nullable: true
          description: Optional description
        parentId:
          type: integer
          nullable: true
          description: Parent category ID for hierarchy
        sortOrder:
          type: integer
          default: 0
          description: Display order within parent
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
    ReportCategoryCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        parentId:
          type: integer
          nullable: true
        sortOrder:
          type: integer
          default: 0
      required:
        - name

    # --- Report Template ---
    ReportTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
          nullable: true
        code:
          type: string
          maxLength: 50
          pattern: '^[a-z0-9-]+$'
          description: URL-safe identifier
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: '#/components/schemas/ReportCategory'
        version:
          type: integer
        isPublished:
          type: boolean
        isSystem:
          type: boolean
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
          description: URL to thumbnail image
        createdBy:
          type: integer
        createdByName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - code
...
    ReportTemplateCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
          nullable: true
        code:
          type: string
          maxLength: 50
          pattern: '^[a-z0-9-]+$'
        categoryId:
          type: integer
          nullable: true
        definition:
          type: string
          description: Initial DevExpress report XML (optional, empty report created if not provided)
      required:
        - name
        - code
...
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        message:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        data:
          type: object
        message:
          type: string
security:
  - bearerAuth: []
paths:
...
  /reports/categories:
    get:
      tags:
        - Report Categories
      summary: List all report categories
      description: Returns hierarchical list of categories with nested children
      parameters:
        - name: includeInactive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ReportCategory'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Report Categories
      summary: Create a new category
      description: Requires admin or manager role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCategoryCreate'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportCategory'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
  /reports/categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags:
        - Report Categories
      summary: Get category by ID
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportCategory'
        '404':
          description: Category not found

    put:
      tags:
        - Report Categories
      summary: Update category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCategoryCreate'
      responses:
        '200':
          description: Category updated
        '400':
          description: Validation error
        '404':
          description: Category not found

    delete:
      tags:
        - Report Categories
      summary: Delete category (soft delete)
      description: Sets isActive to false. Cannot delete if category has templates.
      responses:
        '200':
          description: Category deleted
        '400':
          description: Cannot delete - category has templates
        '404':
          description: Category not found

  # ===========================================
  # NEXT.JS API - REPORT TEMPLATES
  # ===========================================
  /reports/templates:
    get:
      tags:
        - Report Templates
      summary: List report templates
      description: Returns templates the user has permission to view
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: isPublished
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
          description: Search in name and description
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated list of templates
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginatedResponse'
                          - type: object
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/ReportTemplate'

    post:
      tags:
        - Report Templates
      summary: Create a new report template
      description: Requires design permission. Creates draft template.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportTemplateCreate'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportTemplate'
        '400':
          description: Validation error (e.g., duplicate code)
        '403':
          description: Forbidden - no design permission
...